Implement pbrt infrastructure
Reimplement the existing raycaster using the design as described in
pbrt. This lays the foundation for easy extension of functionality later
on. The performance dropped down quite a bit in contrast with the CPU
triangle renderer. But we have glorious camera movement support in
return (using the arrow keys). Furthermore, the OBJ file given through
the command line is now actually displayed instead of two hardcoded
triangles.